Reorganize vue templates.
