+++ resources/js/Partials/Materials/Colors/ColorsTable.vue
