+++ resources/js/Partials/Materials/Fabrics/FabricsTable.vue
