+++ resources/js/Partials/Materials/Inks/InksTable.vue
