+++ resources/js/Partials/Materials/Printers/PrinterColorsTable.vue
